Refactor Recorder and adapt generate_results
For Qubes OS we need to add a list of packages.
